feat: split out functions for better overview
